+++ src/interface/response/GetArticlesResponseInterface.ts
